copperplate/registry: introduce registry daemon
The registry support is managing a FUSE-based filesystem for exporting
the real-time objects created by the Xenomai core on behalf of the
application (e.g. semaphores, tasks, heaps and so on).

The hierarchy of the Xenomai registry is organized as follows:

/mount-point            /* registry fs root, defaults to /mnt/xenomai */
    /session            /* shared session name or "anon" */
        /pid            /* application pid */
           /skin        /* interface name: alchemy/vxworks/psos/... */
              /family   /* object kind (task, semaphore, ...) */
                 { exported objects... }
        /system         /* session-wide information */

Each leaf entry under a session hierarchy is normally viewable, for
retrieving the information attached to the corresponding object, such
as its state, and/or value. There can be multiple sessions hosted
under a single registry mount point. Session-less application
processes are grouped under the special "anon" hierarchy.

The /system hierarchy provides information about the current state of
the Xenomai core, transversely to all processes which belong to the
parent session. Typically, the status of all threads and heaps created
by the session can be retrieved.

The registry daemon is a companion tool managing exactly one registry
mount point, which is specified by the --root option on the command
line. The daemon monitors FUSE client mounts from applications,
removing stale process mount points as the latter come and go. It is
also in charge of exporting the /system hierarchy for each registry
mount point.

The daemon is automatically spawned by the registry support code as
required. There is normally no action required from users for managing
it.

NOTE: the implementation aims at unifying the registry support for
both Cobalt and Mercury cores. Currently, the registry fs is populated
only partially. More objects will be exported in future releases,
converging to a common hierarchy between Cobalt and Mercury.

+++ lib/vxworks/taskLib.c
@@ -188,7 +188,7 @@ static inline char *task_decode_status(struct wind_task *task, char *buf)
 	return buf + 1;
 }
 
-static size_t task_registry_read(struct fsobj *fsobj,
+static ssize_t task_registry_read(struct fsobj *fsobj,
 				 char *buf, size_t size, off_t offset)
 {
 	struct wind_task *task;
@@ -202,7 +202,7 @@ static size_t task_registry_read(struct fsobj *fsobj,
 	len += sprintf(buf + len, "priority   = %d\n", wind_task_get_priority(task));
 	len += sprintf(buf + len, "lock_depth = %d\n", threadobj_get_lockdepth(&task->thobj));
 
-	return len;
+	return (ssize_t)len;
 }
 
 static struct registry_operations registry_ops = {
